% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adstock}
\alias{adstock}
\title{Adstock transformation function.}
\usage{
adstock(x, rate = 0)
}
\arguments{
\item{x}{An ordered numeric none-na vector}

\item{rate}{A decimal between 0-1}
}
\description{
Computes an adstock transformed variable
}
\details{
Transforms variable 'x' into adstock version with rate 'rate'.
}
\examples{
adstock(x=c(10, 14, 30, 32, 0, 0, 0, 10), rate=.35)
}
\author{
Gabriel Mohanna
}
\seealso{
\url{http://analyticsartist.wordpress.com/2013/11/02/calculating-adstock-effect/}
}

