% Generated by roxygen2 (4.0.1): do not edit by hand
\name{completeParameters}
\alias{completeParameters}
\title{Create a complete set of parameters}
\usage{
completeParameters(indp.vars, some.vars, some.vars.name, var.init)
}
\arguments{
\item{indp.vars}{A character list of variable names}

\item{some.vars}{A vector (numeric or character)}

\item{some.vars.name}{A string indicating the name of 'some.vars'}

\item{var.init}{Intitial values to completed for all 'indp.vars' not in 'some.vars'}
}
\description{
Some docs
}
\author{
Gabriel Mohanna
}

