% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createInTrainFolds}
\alias{createInTrainFolds}
\title{Training observations selection.}
\usage{
createInTrainFolds(data, k = 10, type = "bool")
}
\arguments{
\item{data}{A data frame}

\item{k}{An integer for the number of folds}

\item{type}{A character indicating if indicator of inTrain is boolean. Anything else is will produce 1 or 0.}
}
\description{
Sequentially divides the set of observations into k groups, or folds
}
\details{
Observations are divided sequentially into set of k groups, or folds, of approximately equal size. The first fold is treated as a validation set and the other k-1 folds are used for training. This process is repeated for each of the k folds.
}
\examples{
createInTrainFolds(data=iris)
createInTrainFolds(data=iris, 5)
}
\author{
Gabriel Mohanna
}

