% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createInTrainSamples}
\alias{createInTrainSamples}
\title{Training observations selection.}
\usage{
createInTrainSamples(data, times = 1, p = 0.5, type = "bool")
}
\arguments{
\item{data}{A data frame}

\item{p}{The percentage of data that goes to training. This can be a constact or a function.}

\item{times}{The number of partitions to create}

\item{type}{A character indicating if indicator of inTrain is boolean. Anything else is will produce 1 or 0.}
}
\description{
Randomly assigns each observation into training set without replacement
}
\details{
Creates a series of training observations that are randomly selected from the 'data' with probability 'p' without replacement.
}
\examples{
createInTrainSamples(data=iris)
createInTrainSamples(data=iris, p=.7)
createInTrainSamples(data=iris, times=20)
createInTrainSamples(data=iris, times=20, p=.7)
}
\author{
Gabriel Mohanna
}

