% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createInTrainTimeSlices}
\alias{createInTrainTimeSlices}
\title{Observations Splitting Functions}
\usage{
createInTrainTimeSlices(data, initialWindow, horizon = 1, fixedWindow = T)
}
\arguments{
\item{data}{A data frame}

\item{initialWindow}{Something}

\item{horizon}{Something}

\item{fixedWindow}{Something}
}
\description{
Not yet implemented
}
\details{
<empty> for now
}
\author{
Gabriel Mohanna
}

