% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_shortfall.R
\name{es_cornishfisher}
\alias{es_cornishfisher}
\title{Expected Shortall using the Cornish Fisher adjustment for non-normality}
\usage{
es_cornishfisher(mu, sigma, skew, kurt, conf = 0.95)
}
\arguments{
\item{mu}{mean daily P/L data}

\item{sigma}{standard deviation of daily P/L data}

\item{skew}{skewness}

\item{kurt}{kurtosis}

\item{conf}{the confidence level (double)}

\item{holding}{the holding period in days (double)}
}
\value{
ES measure (numeric)
}
\description{
Function estimates the ES for near normal P/L using the  Cornish Fisher
adjustment for non-normality for specified confidence level.
}
