% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_shortfall.R
\name{es_historical}
\alias{es_historical}
\title{Expected Shortfall using the historical simulation approach}
\usage{
es_historical(x, conf = 0.95)
}
\arguments{
\item{x}{P/L data}

\item{conf}{confidence level}
}
\value{
ES measure (numeric)
}
\description{
Estimates the ES by averaging the losses in the tail. Holding time
is inferred by the frequency of the data.
}
