% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_shortfall.R
\name{es_lognormal}
\alias{es_lognormal}
\title{Expected Shortfall for normally distributed geometric returns}
\usage{
es_lognormal(mu, sigma, investment, conf = 0.95, holding = 1)
}
\arguments{
\item{mu}{mean daily P/L data}

\item{sigma}{standard deviation of daily P/L data}

\item{investment}{the size of investment}

\item{conf}{the confidence level (double)}

\item{holding}{the holding period in days (double)}
}
\value{
ES measure (numeric)
}
\description{
This function estimates the ES of P/L data assuming geometric returns
are normally distributed, for specified confidence level and holding period.
It does so by taking an average of the VaRs in the tail of the distribution.
}
