% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_shortfall.R
\name{es_normal}
\alias{es_normal}
\title{Expected Shortfall for normally distributed P/L}
\usage{
es_normal(mu, sigma, conf = 0.95, holding = 1)
}
\arguments{
\item{mu}{mean daily P/L data}

\item{sigma}{standard deviation of daily P/L data}

\item{conf}{the confidence level (double)}

\item{holding}{the holding period in days (double)}
}
\value{
ES measure (numeric)
}
\description{
This function estimates the ES of a portfolio assuming P/L is
normally distributed for a specified confidence level and holding period.
}
