% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightings.R
\name{EWMA_corr}
\alias{EWMA_corr}
\title{Exponentially Weighted Moving Average correlation between
two series of returns}
\usage{
EWMA_corr(x, y, decay = 0.85)
}
\arguments{
\item{x}{y  P/L data}

\item{decay}{decay factor (0 < x < 1)}
}
\value{
EWMA correlation
}
\description{
Exponentially Weighted Moving Average correlation between
two series of returns
}
