% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightings.R
\name{EWMA_cov}
\alias{EWMA_cov}
\title{Exponentially Weighted Moving Average covariance between
two series of returns}
\usage{
EWMA_cov(x, y, decay = 0.85)
}
\arguments{
\item{x, }{y  P/L data}

\item{decay}{decay factor (0 < x < 1)}
}
\value{
EWMA covariance (numeric)
}
\description{
Exponentially Weighted Moving Average covariance between
two series of returns
}
