% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightings.R
\name{EWMA_vol}
\alias{EWMA_vol}
\title{TODO: Filtered Historical Simulation (pg. 69 4.4.3)
Exponentially Weigthed Moving Average for estimating volatility}
\usage{
EWMA_vol(x, decay = 0.85)
}
\arguments{
\item{x}{P/L data}

\item{decay}{decay factor (0 < x < 1)}
}
\value{
EWMA volatilities (numeric)
}
\description{
Estimate volatilities of time series of returns using an EWMA
}
