% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_shortfall.R
\name{expected_shortfall}
\alias{expected_shortfall}
\title{Expected Shortfall}
\usage{
expected_shortfall(x)
}
\description{
Estimate various parametric and non-parametric expected shortfall
measures
}
\examples{
# Compute 5\% ES for the corporate portfolio
portfolio_returns \%>\%
  filter(portfolio == "corporate") \%>\%
  expected_shortfall(
    returns,
    alpha = 0.05,
    method = "historical simulation",
    weighting = weight_exp(lambda = 0.9)
  )

# Compute 1\% ES for each portfolio
portfolio_returns \%>\%
  group_by(portfolio) \%>\%
  expected_shortfall(
    returns,
    alpha = 0.05,
    method = "Cornish Fisher"
  )
}
