% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R
\name{formatted_data}
\alias{formatted_data}
\title{Creating Formatted Data}
\usage{
formatted_data(var_name, format)
}
\arguments{
\item{var_name}{The name of the variable chosen}

\item{format}{Whether or not to format the variable (i.e. percents, dollar
amounts, commas)}
}
\description{
Creating Formatted Data
}
\examples{
mmR:::formatted_data("mcv2", TRUE)
}
