% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Summary statistics of a variable}
\usage{
generate_summary_stats(x)
}
\arguments{
\item{x}{Variable of the form of a vector}
}
\value{
A tibble of columns for the variable, where each column is a different measure
}
\description{
Helper function that calculates summary statistics based on the type of variable
}
\examples{
# summary statistics for the region variable
summ_stats(load_data()$region)
}
