% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getModeledData}
\alias{getModeledData}
\title{Returns data used in optimization}
\usage{
getModeledData(nlsObject)
}
\arguments{
\item{nlsObject}{An nls object.}
}
\description{
Returns data used in optimization without inTrain or intercept
}
\author{
Gabriel Mohanna
}

