% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getModelErrors}
\alias{getModelErrors}
\title{Calculates Model Error}
\usage{
getModelErrors(MMR.object, type = "MSE", training = T)
}
\arguments{
\item{MMR.object}{An MMR.Optim object for which extraction of models error is meaningful}

\item{type}{A character of {"MSE", "MAPE"}}

\item{training}{A logical to indicate inTrain vs. inTest}
}
\description{
Some description
}
\author{
Gabriel Mohanna
}

