% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_multiplication.R
\name{mm}
\alias{mm}
\alias{\%mm\%}
\title{Matrix Multiplication}
\usage{
mm(x, y, return_matrix)

"\%mm\%"(x, y, return_matrix)
}
\arguments{
\item{x}{A numeric matrix or vector}

\item{y}{A numeric matrix or vector}

\item{return_matrix}{Defaults to returning a data.frame (FALSE). Set to TRUE to return a matrix}
}
\value{
The matrix product as a data.frame or matrix

The matrix product as a data.frame or matrix
}
\description{
Multiplies two matricies. A wrapper around `matmalt()` (i.e. `%*%`) that handles
    `data.frame`s

Multiplies two matricies. A wrapper around `matmalt()` (i.e. `%*%`) that handles
    `data.frame`s
}
\examples{

x <- data.frame(a=c(1,2,3), b=c(5,6,7))
y <- c(2,2)
mm(x, y)


x <- data.frame(a=c(1,2,3), b=c(5,6,7))
y <- c(2,2)
x \%mm\% y

}
