% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Adstock.Plot}
\alias{MMR.Adstock.Plot}
\title{Adstock Rates Plot}
\usage{
MMR.Adstock.Plot(object, Param, y)
}
\arguments{
\item{object}{An MMR.Optim object for which extraction of model coeeficients is meaningful}

\item{Param}{A character vector or one or more variables on which adstock is to be plotted}

\item{y}{A character of {"mean", "MAPE", "MSE", "RMSE"}}
}
\description{
Some description
}
\author{
Gabriel Mohanna
}

