% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Coef}
\alias{MMR.Coef}
\title{Prints Coefficient Estimates for MMR.Optim}
\usage{
MMR.Coef(object)
}
\arguments{
\item{object}{An MMR.Optim object for which extraction of model coeeficients is meaningful}
}
\description{
Summarize the coefficient values for MMR.Optim
}
\author{
Gabriel Mohanna
}

