% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Contributions}
\alias{MMR.Contributions}
\title{Contribution Calculations}
\usage{
MMR.Contributions(MMR.object, combine = TRUE, combineMethod = "MSE")
}
\arguments{
\item{object}{An MMR.Optim object for which prediction is desired}

\item{combine}{A logical to if all predictions to be combined}

\item{combineMethod}{A character of {"mean", "MAPE", "MSE", "RMSE"}}
}
\description{
Calculates the contribution
}
\author{
Gabriel Mohanna
}

