% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Get.Trials}
\alias{MMR.Get.Trials}
\title{Trials Generations}
\usage{
MMR.Get.Trials(data, cvControl = cv.control("none"))
}
\arguments{
\item{data}{A data frame}

\item{cvControl}{A list of cross-validation selection parameters.}
}
\description{
Generates a list of inTrain tials
}
\author{
Gabriel Mohanna
}

