% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Optim.Formula}
\alias{MMR.Optim.Formula}
\title{Write Optim Formula}
\usage{
MMR.Optim.Formula(dep.var, fixed = character(), fixed.adstock = rep(0,
  length(fixed)), random = character(), transform = character())
}
\arguments{
\item{dep.var}{A numeric vector}

\item{fixed}{A character vector of variable names that won't be adstocked}

\item{fixed.adstocked}{A numeric vector same length as \code{fixed}}

\item{random}{A character vector of variable names to be adstocked}
}
\description{
Writes the inTrain formula such as "sales ~ inTrain * (Trend + ad1)
}
\author{
Gabriel Mohanna
}

