% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Optim}
\alias{MMR.Optim}
\title{Nonlinear Least Squares Parameter Estimation}
\usage{
MMR.Optim(dep.var, indp.vars = character(), min.adstock, max.adstock,
  keep.vars = character(), transform = character(), data,
  algorithm = "LM", cvTrials, indp.try = length(indp.vars) -
  length(keep.vars))
}
\arguments{
\item{dep.var}{Dependent variable}

\item{indp.vars}{List of variable names to be tried}

\item{min.adstock}{An optional list as a numeric vector of minimum adstock for each \code{indep.vars}}

\item{max.adstock}{An optional list as a numeric vector of maximum adstock for each \code{indep.vars}}

\item{keep.vars}{An optional list of variable names that will be kept in each model. \code{keep.vars} must be part of \code{indep.vars}}

\item{transform}{TBD}

\item{data}{A dataset from in which to evaluate the \code{indp.vars}}

\item{algorithm}{Character string specifying the algorithm to use. Only "LM" based on Levenberg-Marquardt and "port" algorithms are implented.}

\item{cvTrials}{A list of logicals to indicate which variable are part of training data. MMR.Optim will repeat for every elements of \code{cvTrials}}

\item{indp.vars.try}{Number of variables randomly sample as independant variables in each \code{cvTrials} trials}
}
\description{
Determines nonlinear least-squares estimates marketing-mix model
}
\details{
MMR.Optim determines the nonlinear least-squares estimates of marketing-mix model using optimization techniques such as nls & nlsLM.
The current functionality, for now, focuses on estimating the adstock effect with future plans to also determine dimensioning returns using power transformation or something alike.
}
\author{
Gabriel Mohanna
}

