% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Predict}
\alias{MMR.Predict}
\title{Model Prediction}
\usage{
MMR.Predict(MMR.object, newdata, combine = TRUE, combineMethod = "MSE",
  combineTest = TRUE)
}
\arguments{
\item{object}{An MMR.Optim object for which prediction is desired}

\item{newdata}{something}

\item{combine}{A logical to indicate if all predictions to be combined}

\item{combineMethod}{A character of {"mean", "MAPE", "MSE", "RMSE"}}

\item{combineTest}{A logical to indicate to use inTest observations}
}
\description{
Predicts
}
\author{
Gabriel Mohanna
}

