% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Transform}
\alias{MMR.Transform}
\title{MMR transformation function.}
\usage{
MMR.Transform(x, adstock = 0, power = 1, lag = 0)
}
\arguments{
\item{x}{An ordered numeric vector}

\item{adstock}{A decimal between 0-1}

\item{power}{A decimal between 0-1}

\item{lag}{An positive integer}
}
\description{
Function Not Yet Done. Transforms advertising into adstock, power and lag version.
}
\details{
Transforms advertising into adstock version
}
\author{
Gabriel Mohanna
}

