% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Leaflet Plot}
\usage{
plot_leaflet(var, low, high, format = TRUE)
}
\arguments{
\item{var}{The variable to be plotted (mcv2, GDP_per_capita,
measles_incidence_rate_per_1000000_total_population)}

\item{low}{The color for the low end of the scale, as a string}

\item{high}{The color for the high end of the scale, as a string}

\item{format}{Whether or not to format the variable (i.e. percents, dollar
amounts, commas). Default is TRUE}
}
\value{
An interactive leaflet plot
}
\description{
Leaflet Plot
}
\examples{
plot_leaflet(var = mcv2, low = "#ecf0f1", high = "#e74c3c")
plot_leaflet(var = GDP_per_capita, low = "#ecf0f1", high = "#1ccdaa")
plot_leaflet(var = measles_incidence_rate_per_1000000_total_population,
low = "#ecf0f1", high = "#f4a324", format = FALSE)
}
