% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.ads}
\alias{sim.ads}
\title{Advertising Data Simulation}
\usage{
sim.ads(weeks, trend = T, season = T, ads.count = 0, sequantial = T,
  adstock.max = 0, ads.cor = 0, ads.noise = 0)
}
\arguments{
\item{weeks}{Number of weeks, or observations}

\item{trend}{An indicator of a trend sequance}

\item{season}{An indicator of seasonal effect}

\item{ads.count}{Count of advertising medium or campaigns}

\item{ads.cor}{Additial ads that are correlated with 'ads.count'}

\item{ads.noise}{A count of non-related ads with no effects}

\item{adstock.max}{The maximum adstock rate to give to each 'ads.count'}
}
\description{
Creates a random advertising data
}
\details{
Simulates a randomly generated advertising data
}
\examples{
# Simulate 2 years of data with trend, seasonality and 3 different ads.
sim.ads(weeks=104, trend=TRUE, season=TRUE, ads.count=3)

# Simulate 2 years of data with 3 different ads and 1 unrelated variable.
sim.ads(weeks=104, trend=FALSE, season=FALSE, ads.count=3, ads.noise=1)
}
\author{
Gabriel Mohanna
}

