% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{summ_stats_by_region}
\alias{summ_stats_by_region}
\title{Summary statistics table}
\usage{
summ_stats_by_region(r, summ_param, number = 5)
}
\arguments{
\item{r}{Region of interest}

\item{summ_param}{Some parameter of interest}

\item{number}{Number of countries to include}
}
\value{
A gt table
}
\description{
Makes a gt table for top n countries for a region for the specified parameter.
This is similar to dashboard table on Mongolia panel.
}
