% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Summary statistics for a specified variable aggregated by region or country}
\usage{
summarize_stats(df, col = measles_lab_confirmed, mode = region)
}
\arguments{
\item{df}{Data frame}

\item{col}{Some column of interest to summarize by}

\item{mode}{Mode to aggregate by}
}
\value{
A tibble
}
\description{
Helper function used to calculate summary statistics for a specified variable
}
