% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{summarize_stats_by_region}
\alias{summarize_stats_by_region}
\title{Summary statistics table}
\usage{
summarize_stats_by_region(
  r = "Africa",
  summ_param = "GDP_per_capita",
  number = 5
)
}
\arguments{
\item{r}{Region of interest}

\item{summ_param}{Some parameter of interest}

\item{number}{Number of countries to include}
}
\value{
A gt table of formatted summary statistics for a specified variable for a region
}
\description{
Makes a gt table for top n countries for a region for the specified parameter.
This is similar to dashboard table on Mongolia panel.
}
\examples{
# Generate a summary table for the top 5 African countries based on country's GDP
summarize_stats_by_region(r = "Africa", summ_param = "GDP_per_capita")
}
