% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtests.R
\name{test_jb}
\alias{test_jb}
\title{Christoffersen Conditional Backtesting Approach}
\usage{
test_jb()
}
\arguments{
\item{n}{the number of observations}

\item{skew}{the sample skewness}

\item{kurt}{the sample kurtosis}
}
\description{
Jarque Bera Backtest
}
