% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtests.R
\name{test_kupiec}
\alias{test_kupiec}
\title{Binomial Test (Kupiec Test/Frequency-of-tail-losses Test)}
\usage{
test_kupiec(x, n, conf)
}
\arguments{
\item{x}{the number of tail losses}

\item{n}{the number of observations}

\item{conf}{the level of confidence}
}
\value{
probability model is correct (numeric)
}
\description{
Kupiec test carries out the binomial backtest for a VaR
risk measurement model, for specified VaR confidenve level,
for a one sided hypothesis test.
}
