% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trialContrib}
\alias{trialContrib}
\title{Contribution Calculations for One Trial}
\usage{
trialContrib(trial)
}
\arguments{
\item{object}{An MMR.Optim object for which prediction is desired}
}
\description{
Calculates the contribution for one trial
}
\author{
Gabriel Mohanna
}

