% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R
\name{validate_color}
\alias{validate_color}
\title{Color Checker}
\usage{
validate_color(color)
}
\arguments{
\item{color}{The color input to be checked}
}
\description{
Color Checker
}
\examples{
mmR:::validate_color("#ecf0f1")
}
