% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{validate_region}
\alias{validate_region}
\title{Region Name checker}
\usage{
validate_region(region_name)
}
\arguments{
\item{region_name}{Region to checked}
}
\description{
Helper function to validate region names
}
