% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_model.R
\name{validate_v}
\alias{validate_v}
\title{Variable Name checker}
\usage{
validate_v(variable_name)
}
\arguments{
\item{variable_name}{Variable to checked}
}
\description{
Helper function to validate variables for modeling
}
