% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R
\name{validate_var}
\alias{validate_var}
\title{Variable Name Checker}
\usage{
validate_var(var_name)
}
\arguments{
\item{var_name}{The name of the variable to checked}
}
\description{
Variable Name Checker
}
\examples{
mmR:::validate_var("mcv2")
}
