% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats_by_region.R
\name{validate_variable}
\alias{validate_variable}
\title{Variable Name checker}
\usage{
validate_variable(variable_name)
}
\arguments{
\item{variable_name}{Variable to checked}
}
\description{
Helper function to validate variables
}
