% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_at_risk.R
\name{value_at_risk}
\alias{value_at_risk}
\title{Value at Risk}
\usage{
value_at_risk(x)
}
\arguments{
\item{data}{a dataframe of returns}

\item{...}{unquoted column names to use in computation}

\item{alpha}{1 - confidence level}

\item{method}{the method for computing VAR}

\item{weighting}{a function for weighting returns}

\item{params}{a list of named parameters for specific methods}
}
\description{
Estimate a various parameteric and non-parametric value at risk
measures
}
\examples{
# compute 5\% VaR for the corporate portfolio
portfolio_returns \%>\%
  filter(portfolio == "corporate") \%>\%
  value_at_risk(
    returns,
    alpha = 0.05,
    method = "historical simulation"
  )

# Compute 1\% VaR for each portfolio
portfolio_returns \%>\%
  group_by(portfolio) \%>\%
  value_at_risk(
    returns,
    alpha = 0.01,
    method = "historical simulation",
    weighting = weight_vol(decay = 0.01)
  )
}
