% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_at_risk.R
\name{var_historical}
\alias{var_historical}
\title{Historical Value at Risk}
\usage{
var_historical(x, conf = 0.95)
}
\arguments{
\item{x}{a vector of P/L data}

\item{conf}{the confidence level}
}
\value{
VaR measure (numeric)
}
\description{
Estimate the VaR of a portfolio using the historical simulation approach
for specified confidence levels. The holding period is implied by the P/L
data passed to the function. This function uses the quantile, rather than
the method used in Measuring Market Risk - that is approximating the pdf
by drawing lines through the midpoints of a histogram.
}
