% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_at_risk.R
\name{var_lognormal}
\alias{var_lognormal}
\title{Value at Risk for lognormally distributed P/L}
\usage{
var_lognormal(mu, sigma, investment, conf = 0.95, holding = 1)
}
\arguments{
\item{mu}{mean daily P/L}

\item{sigma}{standard deviation of daily P/L}

\item{investment}{the size of investment}

\item{conf}{the confidence level}

\item{holding}{the holding period in days}
}
\value{
VaR measure (numeric)
}
\description{
Estimate the VaR of a portfolio assuming that geometric returns are
normally distributed, for specified confidence level and holding
period.
}
