% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_at_risk.R
\name{var_normal}
\alias{var_normal}
\title{Value At Risk for normally distributed P/L}
\usage{
var_normal(mu, sigma, conf = 0.95, holding = 1)
}
\arguments{
\item{mu}{mean daily P/L}

\item{sigma}{standard deviation of daily P/L}

\item{conf}{the confidence level}

\item{holding}{the holding period in days}
}
\value{
VaR measure (numeric)
}
\description{
Estimate the VaR of a portfolio assuming P/L is =
normally distributed for specified confidence level and holding period.
}
