% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightings.R
\name{weight_by_age}
\alias{weight_by_age}
\title{Age Weighting function}
\usage{
weight_by_age(x, decay = 0.9)
}
\arguments{
\item{x}{P/L data}

\item{decay}{decay factor (0 < x < 1)}
}
\value{
P/L data weighted by age (numeric)
}
\description{
Weight P/L data according to age, using an exponentially decreasing
factor.
}
