% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightings.R
\name{weight_by_vol}
\alias{weight_by_vol}
\title{Volatility Weighting function}
\usage{
weight_by_vol(x, decay)
}
\arguments{
\item{x}{P/L data}

\item{decay}{decay factor (0 < x < 1)}
}
\value{
P/L data weighted by volatility (numeric)
}
\description{
Weight P/L data according to past volatility. This function uses
\code{EWMA_vol} to compute volatility. Proposed by Hull and White (1998).
}
