% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Split a string}
\usage{
load_data(
  link =
    "https://raw.githubusercontent.com/rfordatascience/tidytuesday/main/data/2025/2025-06-24/cases_year.csv"
)
}
\arguments{
\item{string}{A character vector with, at most, one element containing the link to the csv.}
}
\value{
The loaded dataset in the form of a dataframe.
}
\description{
Split a string
}
\examples{
load_data()

y <- 'https://raw.githubusercontent.com/rfordatascience/tidytuesday/main/data/2025/2025-06-24/cases_year.csv'
str_split_one(y)

}
