% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Optim.Formula}
\alias{MMR.Optim.Formula}
\title{Write Optim Formula}
\usage{
MMR.Optim.Formula(dep.var, coef.vars, adstock.min, adstock.max, transform.vars,
  transform.min, transform.max)
}
\arguments{
\item{dep.var}{A numeric vector}

\item{coef.vars}{A character vector}

\item{adstock.min}{A numeric vector}

\item{adstock.max}{A numeric vector}

\item{transform.vars}{A character vector}

\item{transform.min}{A numeric vector}

\item{transform.max}{A numeric vector}
}
\description{
Writes the inTrain formula such as "sales ~ inTrain * (Trend + ad1)
}
\author{
Gabriel Mohanna
}

