% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MMR.Optim.InTrain}
\alias{MMR.Optim.InTrain}
\title{Nonlinear Least Squares Parameter Estimation}
\usage{
MMR.Optim.InTrain(dep.var, indp.vars, min.adstock, max.adstock, keep.vars,
  transform, data, algorithm, inTrain, indp.try)
}
\arguments{
\item{dep.var}{Dependent variable}

\item{indp.vars}{List of variable names to be tried}

\item{min.adstock}{An optional list as a numeric vector of minimum adstock for each \code{indp.vars}}

\item{max.adstock}{An optional list as a numeric vector of maximum adstock for each \code{indp.vars}}

\item{keep.vars}{An optional list of variable names that will be kept in each model. \code{keep.vars} must be part of \code{indep.vars}}

\item{transform}{TBD}

\item{data}{A dataset from in which to evaluate the \code{indp.vars}}

\item{algorithm}{The algorithm to use {"LM", "port"}}

\item{inTrain}{A logical list indicating which observations are used for training}

\item{indp.try}{A positive decimal to indicate the probability of including all random.vars}
}
\description{
Determines nonlinear least-squares estimates marketing-mix model for a single trial.
}
\examples{
\donttest{MMR.Optim.InTrain("sales", c("Trend", "Season"), c("ad1", "ad2", "ad3"), ad.data, sample(c(T, F), nrow(ad.data)))}
}

