% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_model.R
\name{model_form}
\alias{model_form}
\title{Modeling function}
\usage{
model_form(variables)
}
\arguments{
\item{variables}{Vector of explanatory variables.}
}
\value{
Model of class \code{gls}.
}
\description{
Makes a gls model using number of measles cases per million population as the dependent variable.
Since the model uses yearly data, the model utilizes an autocorrelation structure for the errors using year as the covariate.
}
\examples{
# Make a model using `GDP_per_capita` and `mcv2` as explanatory variables
model_form(c("GDP_per_capita", "mcv2"))
}
