% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Leaflet Plot}
\usage{
plot_leaflet(var, low, high, format, normalize = FALSE)
}
\arguments{
\item{var}{The variable to be plotted (mcv2, GDP_per_capita,
measles_incidence_rate_per_1000000_total_population)}

\item{low}{The color for the low end of the scale, as a string}

\item{high}{The color for the high end of the scale, as a string}

\item{format}{How the parameter should be formatted ("percent", "dollar",
"comma")}

\item{normalize}{Whether or not the variable should be normalized, should be
TRUE for variables not on a 0-1 scale. Default is FALSE}
}
\value{
An interactive leaflet plot
}
\description{
Leaflet Plot
}
\examples{
plot_leaflet(var = mcv2, low = "#ecf0f1", high = "#e74c3c",
format = "percent")
plot_leaflet(var = GDP_per_capita, low = "#ecf0f1", high = "#1ccdaa",
format = "dollar", normalize = TRUE)
plot_leaflet(var = measles_incidence_rate_per_1000000_total_population,
low = "#ecf0f1", high = "#f4a324", format = "comma", normalize = TRUE)
}
