% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{time_lapse_plot}
\alias{time_lapse_plot}
\title{Generates a time series plot of a given parameter over a specified time range, with optional filtering.}
\usage{
time_lapse_plot(
  parameter = "Measles",
  start_year = "2012",
  end_year = "2024",
  title_ = glue("{parameter} from {start_year} to {end_year}"),
  filepath = NULL,
  filter_vector = NULL,
  palette = c("#001E4D", "#F4A324")
)
}
\arguments{
\item{parameter}{A character string specifying the parameter to be plotted (Default:, "Measles")}

\item{start_year}{A character string specifying the starting year for the time series (Default: "2012")}

\item{end_year}{A character string specifying the ending year for the time series (Default: "2024")}

\item{filepath}{A character string specifying the file path where the generated *.gif file will be saved (Default: "time_lapse.gif")}

\item{filter_vector}{An optional vector of character strings to filter the data by specific categories or groups (Default: NULL)}

\item{title}{A character string specifying the title of the plot (Default: NULL)}
}
\value{
A *.gif file showing the animated time series of the specified parameter over the given time range.
}
\description{
Generates a time series plot of a given parameter over a specified time range, with optional filtering.
}
