% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-leaflet.R, R/summ_stats_by_region.R
\name{validate_var}
\alias{validate_var}
\title{Variable Name Checker}
\usage{
validate_var(var_name)

validate_var(var_name)
}
\arguments{
\item{var_name}{Variable to checked}
}
\description{
Helper function to validate variables
}
\examples{
mmR:::validate_var("mcv2")
}
